speech2text/pyproject.toml
[project]
name = "voice-input"
version = "0.1.0"
description = "Voice-to-text input for Wayland/Linux using OpenAI Whisper"
requires-python = ">=3.11"
dependencies = [
    "pyaudio>=0.2.14",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
